requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
alembic==1.13.1
pysqlite3-binary==0.5.2
pydantic==2.6.1
python-decouple==3.8
python-dotenv==1.0.1
jinja2==3.1.3
transformers==4.37.2
torch==2.2.0
sentencepiece==0.1.99